+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## 1.2.0
 
 * feature: add support for copyWith method.
+* fixed: Data type conversion error, the same key specifies different data types.
 
 ## 1.1.1